Feat: Request 타입 추가
