+++ src/type/apiType.ts
@@ -1,0 +1,11 @@
+export interface Response<Data> {
+  timestamp: string;
+  code: string;
+  message: string;
+  data: Data;
+}
+
+export interface Request<Data, Identifier = null> {
+  id?: Identifier;
+  data: Data;
+}
