--- src/type/authType.ts
@@ -1,10 +1,0 @@
-export interface LoginRequestModel {
-  memberId: string;
-  password: string;
-}
-
-export interface LoginResponseModel {
-  accessToken: string;
-  refreshToken: string;
-  memberName: string;
-}
