--- src/type/responseType.ts
@@ -1,6 +1,0 @@
-export default interface Response<Data> {
-  timestamp: string;
-  code: string;
-  message: string;
-  data: Data;
-}
